对0757192c及其之前的5个提交对PackageManager改动的补充。 PluginPackageManager没有必要继承系统的PackageManager了。 删除无用的CommonPluginPackageManager。
