+++ projects/sdk/core/loader/src/main/kotlin/com/tencent/shadow/core/loader/blocs/LoadPluginBloc.kt
@@ -27,7 +27,7 @@ import com.tencent.shadow.core.load_parameters.LoadParameters
 import com.tencent.shadow.core.loader.exceptions.LoadPluginException
 import com.tencent.shadow.core.loader.infos.PluginParts
 import com.tencent.shadow.core.loader.managers.ComponentManager
-import com.tencent.shadow.core.loader.managers.PluginPackageManager
+import com.tencent.shadow.core.loader.managers.PluginPackageManagerImpl
 import com.tencent.shadow.core.runtime.PluginPartInfo
 import com.tencent.shadow.core.runtime.PluginPartInfoManager
 import com.tencent.shadow.core.runtime.ShadowContext
@@ -101,7 +101,7 @@ object LoadPluginBloc {
             val buildPackageManager = executorService.submit(Callable {
                 val packageInfo = getPackageInfo.get()
                 val hostPackageManager = hostAppContext.packageManager
-                PluginPackageManager(hostPackageManager, packageInfo, allPluginPackageInfo)
+                PluginPackageManagerImpl(hostPackageManager, packageInfo, allPluginPackageInfo)
             })
 
             val buildResources = executorService.submit(Callable {
