+++ projects/sdk/core/runtime/src/main/java/com/tencent/shadow/core/runtime/PluginPackageManager.java
@@ -1,0 +1,20 @@
+package com.tencent.shadow.core.runtime;
+
+import android.content.ComponentName;
+import android.content.pm.ActivityInfo;
+import android.content.pm.ApplicationInfo;
+import android.content.pm.PackageInfo;
+import android.content.pm.ProviderInfo;
+import android.content.pm.VersionedPackage;
+
+public interface PluginPackageManager {
+    ApplicationInfo getApplicationInfo(String packageName, int flags);
+
+    ActivityInfo getActivityInfo(ComponentName component, int flags);
+
+    PackageInfo getPackageInfo(String packageName, int flags);
+
+    PackageInfo getPackageInfo(VersionedPackage versionedPackage, int flags);
+
+    ProviderInfo resolveContentProvider(String name, int flags);
+}
